requirements.txt
asgiref==3.8.1
Django==5.1.3
django-clearcache==1.2.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.1
